@@ -1,13 +1,40 @@
 module github.com/TheCacophonyProject/event-reporter/v3
 
-go 1.15
+go 1.22.3
 
 require (
 	github.com/TheCacophonyProject/go-api v1.0.4
+	github.com/TheCacophonyProject/go-utils v0.1.1
 	github.com/TheCacophonyProject/modemd v1.6.0
 	github.com/alexflint/go-arg v1.4.2
 	github.com/boltdb/bolt v1.3.1
 	github.com/coreos/go-systemd/v22 v22.3.2
 	github.com/godbus/dbus v4.1.0+incompatible
+	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.7.0
+)
+
+require (
+	github.com/TheCacophonyProject/go-config v1.9.0 // indirect
+	github.com/alexflint/go-scalar v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/fsnotify/fsnotify v1.5.1 // indirect
+	github.com/godbus/dbus/v5 v5.0.4 // indirect
+	github.com/gofrs/flock v0.8.1 // indirect
+	github.com/hashicorp/hcl v1.0.0 // indirect
+	github.com/magiconair/properties v1.8.5 // indirect
+	github.com/mitchellh/mapstructure v1.4.2 // indirect
+	github.com/pelletier/go-toml v1.9.4 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/afero v1.6.0 // indirect
+	github.com/spf13/cast v1.4.1 // indirect
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/spf13/viper v1.9.0 // indirect
+	github.com/subosito/gotenv v1.2.0 // indirect
+	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
+	golang.org/x/text v0.3.7 // indirect
+	gopkg.in/ini.v1 v1.64.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
